requirements.txt
feedparser
requests
PyYAML
beautifulsoup4
jinja2
anthropic
feedgenerator
pytz